removed transaction funcs and combined to just SQLite Database Functions
