+++ test-tools/DataGen/src/FileCreation/DataGenerator.py
@@ -12,10 +12,9 @@ from Config.Data_Visualizer import log_data
 
 
 def generate_claim_transactions(num_claims, transaction_funcs, input_date=date.today(), status="Created"):
-    localdb = transaction_funcs.family_db
     npi_funcs = transaction_funcs.npi_funcs
 
-    beneficiaries = localdb.get_random_beneficiary(num_claims)
+    beneficiaries = transaction_funcs.get_random_beneficiary(num_claims)
     transactions = []
 
     for bene in beneficiaries:
