+++ test-tools/DataGen/src/RunGenerator.py
@@ -10,12 +10,11 @@ from Config.Data_Visualizer import log_data, create_md
 from FileCreation.DataGenerator import SponsorDataGenerator, generate_claim_transactions
 from FileCreation.EDIGenerator import EDI834Generator, EDI270Generator, EDI837PGenerator, EDI277CAGenerator, \
     EDI835Generator
-from Repository.DatabaseFactory import get_claim_transaction_backend
 from Repository.DatabaseFactory import get_database_backend
 
 load_dotenv()
 
-transaction_funcs = get_claim_transaction_backend()
+transaction_funcs = get_database_backend()
 
 status_map = {
     "270": ("Created", date.today()),
